images: Configure sudo on base image
My base image had already sudo installed because the official Ubuntu
20.04 image has it, however I was not using it and I used the `su`
command when I needed to run something with root privileges.

Now, I have the need to run quite often several commands with root
privileges for Storj, hence, having sudo is more comfortable and
productive rather than having to log in as root and not having all the
shell configurations that my usual user has.

This commit changes the base image to make sure that sudo is installed
and configure it to be used by the normal user under a password because
I prefer for security reasons not allow to elevate the user privileges
without typing a password.